Filesystem Detail Chart shows incorrect units for storage statistics
Fixes #22.

The storage metrics on the filesystem detail page and targets should be calculated using the number-formatters algorithm. Unfortunately, the old gui will not allow the repo to be included and bundled, so for now I think it makes sense to replace the functionality in misc.js with the logic in number-formatters.

Signed-off-by: Will Johnson <wjohnson@whamcloud.com>